Remove inline dataview fields in favor of frontmatter.
The reason why I decided to use inline dataview fields was because at
the time, frontmatter did not support linking to other notes. This is
not the case anymore.

Although note that linking to more than one note per frontmatter field
is not implemented properly and will be implemented later.